[refactor] : common 디렉토리 내부 클래스 의존하던 클래스들 변경 사항 적용
