+++ src/main/java/com/halo/eventer/domain/notice/controller/NoticeController.java
@@ -2,8 +2,8 @@ package com.halo.eventer.domain.notice.controller;
 
 import com.halo.eventer.domain.notice.dto.*;
 import com.halo.eventer.domain.notice.service.NoticeService;
-import com.halo.eventer.global.common.PagedResponse;
-import com.halo.eventer.global.common.ArticleType;
+import com.halo.eventer.global.common.page.PagedResponse;
+import com.halo.eventer.domain.notice.ArticleType;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
