+++ src/main/java/com/halo/eventer/domain/notice/dto/NoticeResDto.java
@@ -2,7 +2,7 @@ package com.halo.eventer.domain.notice.dto;
 
 import com.halo.eventer.domain.image.dto.ImageDto;
 import com.halo.eventer.domain.notice.Notice;
-import com.halo.eventer.global.common.ArticleType;
+import com.halo.eventer.domain.notice.ArticleType;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
