+++ src/main/java/com/halo/eventer/domain/widget/controller/MiddleWidgetController.java
@@ -1,10 +1,10 @@
 package com.halo.eventer.domain.widget.controller;
 
-import com.halo.eventer.global.common.PagedResponse;
-import com.halo.eventer.domain.widget.dto.WidgetOrderUpdateRequest;
+import com.halo.eventer.global.common.page.PagedResponse;
+import com.halo.eventer.global.common.dto.OrderUpdateRequest;
 import com.halo.eventer.domain.widget.dto.middle_widget.*;
 import com.halo.eventer.domain.widget.service.MiddleWidgetService;
-import com.halo.eventer.global.common.SortOption;
+import com.halo.eventer.global.common.sort.SortOption;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
@@ -55,7 +55,7 @@ public class MiddleWidgetController {
 
   @PatchMapping("/displayOrder")
   public List<MiddleWidgetResDto> updateDisplayOrder(@RequestParam Long festivalId,
-                                                     @RequestBody List<WidgetOrderUpdateRequest> orderRequests) {
+                                                     @RequestBody List<OrderUpdateRequest> orderRequests) {
     return middleWidgetService.updateDisplayOrder(festivalId, orderRequests);
   }
 }
