+++ src/main/java/com/halo/eventer/domain/widget/service/MainWidgetService.java
@@ -3,17 +3,13 @@ package com.halo.eventer.domain.widget.service;
 import com.halo.eventer.domain.festival.Festival;
 import com.halo.eventer.domain.festival.exception.FestivalNotFoundException;
 import com.halo.eventer.domain.festival.repository.FestivalRepository;
-import com.halo.eventer.domain.widget.dto.WidgetOrderUpdateRequest;
-import com.halo.eventer.domain.widget.dto.down_widget.DownWidgetResDto;
 import com.halo.eventer.domain.widget.dto.main_widget.MainWidgetCreateDto;
 import com.halo.eventer.domain.widget.dto.main_widget.MainWidgetResDto;
-import com.halo.eventer.domain.widget.entity.DownWidget;
 import com.halo.eventer.domain.widget.entity.MainWidget;
 import com.halo.eventer.domain.widget.exception.WidgetNotFoundException;
 import com.halo.eventer.domain.widget.repository.DownWidgetRepository;
 import com.halo.eventer.domain.widget.repository.MainWidgetRepository;
-import com.halo.eventer.domain.widget.util.DisplayOrderUtils;
-import com.halo.eventer.global.common.WidgetType;
+import com.halo.eventer.domain.widget.WidgetType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
