+++ src/main/java/com/halo/eventer/domain/widget/service/UpWidgetService.java
@@ -3,7 +3,7 @@ package com.halo.eventer.domain.widget.service;
 import com.halo.eventer.domain.festival.Festival;
 import com.halo.eventer.domain.festival.exception.FestivalNotFoundException;
 import com.halo.eventer.domain.festival.repository.FestivalRepository;
-import com.halo.eventer.global.common.PagedResponse;
+import com.halo.eventer.global.common.page.PagedResponse;
 import com.halo.eventer.domain.widget.dto.up_widget.UpWidgetResDto;
 import com.halo.eventer.domain.widget.exception.WidgetNotFoundException;
 import com.halo.eventer.domain.widget.repository.UpWidgetRepository;
@@ -11,8 +11,8 @@ import com.halo.eventer.domain.widget.dto.up_widget.UpWidgetCreateDto;
 import com.halo.eventer.domain.widget.entity.UpWidget;
 
 import com.halo.eventer.domain.widget.util.WidgetPageHelper;
-import com.halo.eventer.global.common.SortOption;
-import com.halo.eventer.global.common.WidgetType;
+import com.halo.eventer.global.common.sort.SortOption;
+import com.halo.eventer.domain.widget.WidgetType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
